Created Formatter class to follow Single Responsibility Principle
